add SpanExtractorWithSpanWidthEmbedding
