+++ allennlp/modules/span_extractors/span_extractor.py
@@ -1,12 +1,7 @@
-from typing import Optional
+import torch
 from overrides import overrides
 
-import torch
-
-from allennlp.common.checks import ConfigurationError
 from allennlp.common.registrable import Registrable
-from allennlp.modules.token_embedders.embedding import Embedding
-from allennlp.nn import util
 
 
 class SpanExtractor(torch.nn.Module, Registrable):
@@ -19,49 +14,7 @@ class SpanExtractor(torch.nn.Module, Registrable):
     and indices of shape (batch_size, num_spans, 2) and return a tensor of
     shape (batch_size, num_spans, ...), forming some representation of the
     spans.
-
-    # Parameters
-
-    input_dim : `int`, required.
-        The final dimension of the `sequence_tensor`.
-    num_width_embeddings : `int`, optional (default = `None`).
-        Specifies the number of buckets to use when representing
-        span width features.
-    span_width_embedding_dim : `int`, optional (default = `None`).
-        The embedding size for the span_width features.
-    bucket_widths : `bool`, optional (default = `False`).
-        Whether to bucket the span widths into log-space buckets. If `False`,
-        the raw span widths are used.
-
-    # Returns
-
-    span_embeddings : `torch.FloatTensor`.
-        A tensor of shape `(batch_size, num_spans, embedded_span_size)`,
-        where `embedded_span_size` depends on the way spans are represented.
     """
-
-    def __init__(
-        self,
-        input_dim: int,
-        num_width_embeddings: int = None,
-        span_width_embedding_dim: int = None,
-        bucket_widths: bool = False,
-    ) -> None:
-        super().__init__()
-        self._input_dim = input_dim
-        self._num_width_embeddings = num_width_embeddings
-        self._bucket_widths = bucket_widths
-
-        self._span_width_embedding: Optional[Embedding] = None
-        if num_width_embeddings is not None and span_width_embedding_dim is not None:
-            self._span_width_embedding = Embedding(
-                num_embeddings=num_width_embeddings, embedding_dim=span_width_embedding_dim
-            )
-        elif num_width_embeddings is not None or span_width_embedding_dim is not None:
-            raise ConfigurationError(
-                "To use a span width embedding representation, you must"
-                "specify both num_width_embeddings and span_width_embedding_dim."
-            )
 
     @overrides
     def forward(
@@ -76,9 +29,7 @@ class SpanExtractor(torch.nn.Module, Registrable):
         them. Span representation can be computed in many different ways,
         such as concatenation of the start and end spans, attention over the
         vectors contained inside the span, etc.
-
         # Parameters
-
         sequence_tensor : `torch.FloatTensor`, required.
             A tensor of shape (batch_size, sequence_length, embedding_size)
             representing an embedded sequence of words.
@@ -99,42 +50,6 @@ class SpanExtractor(torch.nn.Module, Registrable):
 
         A tensor of shape `(batch_size, num_spans, embedded_span_size)`,
         where `embedded_span_size` depends on the way spans are represented.
-        """
-        # shape (batch_size, num_spans, embedding_dim)
-        span_embeddings = self._embed_spans(
-            sequence_tensor, span_indices, sequence_mask, span_indices_mask
-        )
-        if self._span_width_embedding is not None:
-            # width = end_index - start_index + 1 since `SpanField` use inclusive indices.
-            # But here we do not add 1 beacuse we offen initiate the span width
-            # embedding matrix with `num_width_embeddings = max_span_width`
-            # shape (batch_size, num_spans)
-            widths_minus_one = span_indices[..., 1] - span_indices[..., 0]
-
-            if self._bucket_widths:
-                widths_minus_one = util.bucket_values(
-                    widths_minus_one, num_total_buckets=self._num_width_embeddings  # type: ignore
-                )
-
-            # Embed the span widths and concatenate to the rest of the representations.
-            span_width_embeddings = self._span_width_embedding(widths_minus_one)
-            span_embeddings = torch.cat([span_embeddings, span_width_embeddings], -1)
-
-        if span_indices_mask is not None:
-            # Here we are masking the spans which were originally passed in as padding.
-            return span_embeddings * span_indices_mask.unsqueeze(-1)
-
-        return span_embeddings
-
-    def _embed_spans(
-        self,
-        sequence_tensor: torch.FloatTensor,
-        span_indices: torch.LongTensor,
-        sequence_mask: torch.BoolTensor = None,
-        span_indices_mask: torch.BoolTensor = None,
-    ) -> torch.Tensor:
-        """
-        Returns the span embeddings computed in many different ways.
         """
         raise NotImplementedError
